Add entries to gitignore

add_to_gitignore adds entries to gitignore files

add_to_gitignore(add.to.gitignore = FALSE, cutoff = 99, extension = NULL)

Arguments

add.to.gitignore
Logical to control if files are added to 'gitignore' or just printed on the console.

cutoff
Numeric. Defines the file size (in MB) cutoff used to find files (i.e. only files above the threshold would returned). 99 (MB) is recommended when hosting projects at github as the current file size limit is 100 MB. Default is 99.

extension
Character string to define the file extension of the files to be searched for.

Value

Prints the name of the files matching the searching parameters. If add.to.ignore = TRUE the files matching the search parameters ('cutoff' and/or 'extension') would be added 'gitignore' (a file used by git to exclude files form version control, including adding them to github). Hence this can be used to avoid conflicts when working with large files or just avoid adding non-binary files to remote repositories.

Details

The function can be used large size files h.

References

Araya-Salas, M., Willink, B., Arriaga, A. (2020), sketchy: research compendiums for data analysis in R. R package version 1.0.0.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
data(compendiums)

make_compendium(name = "my_compendium", path = tempdir(),
 format = compendiums$basic$skeleton, force = TRUE)

# must start git monitoring at this point
# add_to_gitignore(add.to.gitignore = TRUE)
}

#> Creating directories ...
#> my_compendium
#> │   
#> ├── data/  
#> │   ├── processed/  
#> │   └── raw/  
#> ├── manuscript/  
#> ├── output/  
#> └── scripts/  
#> Done.
#>