Generate folder structures for research compendiums

git_ignore_files generates the folder structure of a research compendium.

git_ignore_files(name = "research_compendium", path = ".", force = FALSE, format,
comments = NULL, packrat = FALSE, git = FALSE, clone = NULL)

Arguments

name
character string: the research compendium directory name. No special characters should be used. Default is "research_compendium".

path
path to put the package directory in. Default is current directory.

force
Logical controlling whether existing folders with the same name are used for setting the folder structure. The function will never overwrite existing files or folders.

format
A character vector with the names of the folders and subfolders to be included. Take a look at `compendiums$basic` for an example.

comments
A character string with the comments to be added to each folder in the graphical representation of the folder skeleton printed on the console.

packrat
Logical to control if packrat is initialized (packrat::init()) when creating the compendium. Default is FALSE.

git
Logical to control if a git repository is initialized (git2r::init()) when creating the compendium. Default is FALSE.

clone
Path to a directory containing a folder structure to be cloned. Default is NULL. If provided 'format' is ignored. Folders starting with ^\.git|^\.Rproj.user|^\.\.Rcheck will be ignored.

Value

A folder skeleton for a research compendium. In addition the structure of the compendium is printed in the console. If the compendium format includes a "manuscript" folder the function saves a manuscript template in Rmd format inside that folder.

Details

The function takes predefined folder structures to generate the directory skeleton of a research compendium.

References

Araya-Salas, M., Willink, B., Arriaga, A. (2020), sketchy: research compendiums for data analysis in R. R package version 1.0.0.Marwick, B., Boettiger, C., & Mullen, L. (2018). Packaging Data Analytical Work Reproducibly Using R (and Friends). American Statistician, 72(1), 80-88.Alston, J., & Rick, J. (2020). A Beginners Guide to Conducting Reproducible Research.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
data(compendiums)

git_ignore_files(name = "mycompendium", path = tempdir(), format = compendiums$basic$skeleton)
}

#> Creating directories ...
#> Done.
#> mycompendium
#> │   
#> ├── data/  
#> │   ├── processed/  
#> │   └── raw/  
#> ├── manuscript/  
#> ├── output/  
#> └── scripts/