Install and load packages

load_packages installs and loads packages from different repositories.

Usage

load_packages(packages, quite = FALSE, upgrade.deps = FALSE)

Arguments

packages

Character vector with the names of the packages to be installed. The vector names indicate the repositories from which packages will be installed. If no name is included CRAN will be used as the default repository. Available repostories are: 'cran', 'github', 'gitlab', 'bitbucket' and 'bioconductor'. Note that for 'github', 'gitlab' and 'bitbucket' the string must include the user name in the form 'user/package'.

quite

Logical argument to control if package startup messages are printed. Default is FALSE (messages are printed).

upgrade.deps

Logical argument to control if package dependecies are upgraded.Default is FALSE.

Value

No object is returned.

Details

The function install and load packages from different repositories.

References

Araya-Salas, M., Willink, B., Arriaga, A. (2020), sketchy: research compendiums for data analysis in R. R package version 1.0.2.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
load_packages(packages = c("kableExtra", bioconductor = "ggtree", github = "maRce10/Rraven"))
}
#> Installing 1 packages: kableExtra
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> also installing the dependencies ‘colorspace’, ‘selectr’, ‘farver’, ‘labeling’, ‘munsell’, ‘RColorBrewer’, ‘rvest’, ‘scales’, ‘viridisLite’, ‘rstudioapi’, ‘webshot’, ‘svglite’
#> Loading required package: kableExtra
#> Downloading Bioconductor repo https://git.bioconductor.org/packages/ggtree
#> Downloading GitHub repo GuangchuangYu/treeio@HEAD
#> Installing 9 packages: generics, tidyselect, dplyr, Rcpp, yulab.utils, tidyr, lazyeval, ape, tidytree
#> Installing packages into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Running `R CMD build`...
#> * checking for file ‘/tmp/RtmpmwsdUZ/remotes1e63504017d9/YuLab-SMU-treeio-db85803/DESCRIPTION’ ... OK
#> * preparing ‘treeio’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> * building ‘treeio_1.23.0.tar.gz’
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Installing 17 packages: generics, Rcpp, yulab.utils, tidyselect, tidyr, lazyeval, dplyr, ape, isoband, gtable, ggplot2, gridGraphics, tidytree, patchwork, ggplotify, ggfun, aplot
#> Installing packages into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Downloading GitHub repo GuangchuangYu/treeio@HEAD
#> Skipping treeio, it is already being installed
#> Running `R CMD build`...
#> * checking for file ‘/tmp/RtmpmwsdUZ/file1e637198b5bc/DESCRIPTION’ ... OK
#> * preparing ‘ggtree’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> * building ‘ggtree_3.6.2.tar.gz’
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Loading required package: ggtree
#> ggtree v3.6.2 For help: https://yulab-smu.top/treedata-book/
#> 
#> If you use the ggtree package suite in published research, please cite
#> the appropriate paper(s):
#> 
#> Guangchuang Yu, David Smith, Huachen Zhu, Yi Guan, Tommy Tsan-Yuk Lam.
#> ggtree: an R package for visualization and annotation of phylogenetic
#> trees with their covariates and other associated data. Methods in
#> Ecology and Evolution. 2017, 8(1):28-36. doi:10.1111/2041-210X.12628
#> 
#> G Yu. Data Integration, Manipulation and Visualization of Phylogenetic
#> Trees (1st ed.). Chapman and Hall/CRC. 2022. ISBN: 9781032233574
#> 
#> Guangchuang Yu. Using ggtree to visualize data on tree-like structures.
#> Current Protocols in Bioinformatics. 2020, 69:e96. doi:10.1002/cpbi.96
#> 
#> Using github PAT from envvar GITHUB_PAT
#> Downloading GitHub repo maRce10/Rraven@HEAD
#> Installing 15 packages: signal, tuneR, moments, bitops, proxy, bioacoustics, rjson, RCurl, pbapply, monitoR, fftw, dtw, NatureSounds, seewave, warbleR
#> Installing packages into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Running `R CMD build`...
#> * checking for file ‘/tmp/RtmpmwsdUZ/remotes1e633914499c/maRce10-Rraven-61e7b62/DESCRIPTION’ ... OK
#> * preparing ‘Rraven’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> * building ‘Rraven_1.0.14.tar.gz’
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Loading required package: Rraven